requirements.txt
streamlit==1.15.2
Pillow==9.5.0
opencv-python-headless==4.5.5.62
numpy==1.24.2
PyMuPDF==1.22.3
tensorflow==2.12.0